+++ Probability/Empty.cs
@@ -10,7 +10,7 @@ namespace Probability
         private Empty() { }
         public T Sample() => throw new Exception("Cannot sample from empty distribution");
         public IEnumerable<T> Support() => Enumerable.Empty<T>();
-        public int Weight(T t) => 0;
+        public long Weight(T t) => 0;
         double IWeightedDistribution<T>.Weight(T t) => 0;
     }
 }
