+++ Probability/IDiscreteDistribution.cs
@@ -4,6 +4,6 @@ namespace Probability
     public interface IDiscreteDistribution<T> : IWeightedDistribution<T>
     {
         IEnumerable<T> Support();
-        new int Weight(T t);
+        new long Weight(T t);
     }
 }
